dataset_generation/requirements.txt
anthropic>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
pymongo>=4.5.0
bson>=0.5.10